tests/wpt/tests/css/css-anchor-position/anchor-position-multicol-006.html

<!DOCTYPE html>
<title>Tests two OOF anchors in different containing blocks in multicol.</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-pos">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.relpos {
  position: relative;
}
.columns {
  column-count: 3;
  column-fill: auto;
  column-gap: 10px;
  column-width: 100px;
  width: 320px;
  height: 50px;
}
.anchor1 {
  anchor-name: --a1;
  position: absolute;
  width: 10px;
  height: 30px;
  background: orange;
}
.anchor2 {
  anchor-name: --a2;
  position: absolute;
  width: 30px;
  height: 30px;
  background: purple;
}
.target {
  position: absolute;
  width: 5px;
  background: lime;
}
.target1 {
  left: anchor(--a1 left);
  top: anchor(--a1 top);
  height: anchor-size(--a1 height);
}
.target2 {
  left: anchor(--a2 left);
  top: anchor(--a2 top);
  height: anchor-size(--a2 height);
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <div class="spacer" style="height: 10px"></div>
  <div class="relpos">
    <div class="columns">
      <div class="relpos">
        <div class="spacer" style="height: 30px"></div>
        <div class="anchor1"></div>
      </div>
      <div class="spacer" style="height: 70px"></div>
      <div class="relpos">
        <div class="spacer" style="height: 10px"></div>
        <div class="anchor2"></div>
      </div>
      <div class="target target1" data-expected-height=50></div>
      <div class="target target2" data-expected-height=30></div>
    </div>
    <div class="target target1" data-expected-height=50></div>
    <div class="target target2" data-expected-height=30></div>
  </div>
</body>
